+++ app/(views)/images/[thread_id]/@modal/default.tsx
@@ -1,0 +1,3 @@
+export default function Default() {
+  return null
+}
