+++ app/(views)/images/[thread_id]/layout.tsx
@@ -6,15 +6,18 @@ import { ThreadHeader } from '@/components/_v/ThreadHeader'
 
 export default function Layout({
   children,
+  modal,
   params,
 }: {
   children: React.ReactNode
+  modal: React.ReactNode
   params: { thread_id: string }
 }) {
   return (
     <>
       <ThreadHeader thread_id={params.thread_id} />
       {children}
+      {modal}
     </>
   )
 }
